move stack and ip walking code into a separate file
