+++ src/Makefile
@@ -29,7 +29,7 @@ SRCS := \
 	jltypes gf ast builtins module interpreter \
 	alloc dlload sys init task array dump toplevel jl_uv jlapi signal-handling \
 	simplevector APInt-C runtime_intrinsics runtime_ccall \
-	threadgroup threading
+	threadgroup threading stackwalk
 
 ifeq ($(JULIAGC),MARKSWEEP)
 SRCS += gc
