+++ src/julia.expmap
@@ -14,7 +14,6 @@
     iswprint;
     jl_*;
     rec_backtrace;
-    rec_backtrace_ctx;
     julia_*;
     libsupport_init;
     localtime_r;
